+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v 4.4.0 (2019/05/12)
+
+* [NEW] - now playlist can be sorted now (drag & drop)
+* [UPGRADE] - more refined error handling
+
 ## v 4.3.0 (2019/05/08)
 
 * [UPGRADE] - theme colors are now as background in user menu